requirements.txt
Flask
faiss-cpu
sentence-transformers
langchain-groq
PyPDF2
python-dotenv
langchain
googletrans==4.0.0-rc1
youtube-transcript-api==0.6.1
waitress==2.1.2
